requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
snowflake-connector-python==3.10.1
apscheduler==3.10.4
streamlit==1.35.0
plotly==5.22.0
pandas==2.1.4
requests==2.32.2
slowapi==0.1.9
